LSM_2023/05/01
네트워크 지연에 대해서 열심히 고치는 중...
최대한 노력 중

+++ 01Script/LSM_SC/ActorScript/LSM_NexusSC.cs
@@ -18,7 +18,7 @@ public class LSM_NexusSC : LSM_TurretSc
 
 	}
 
-	public override void Damaged(int dam, Vector3 origin, MoonHeader.Team t, GameObject other)
+	public override void Damaged(short dam, Vector3 origin, MoonHeader.Team t, GameObject other)
 	{
 		if (t == this.stats.actorHealth.team || !PhotonNetwork.IsMasterClient)
 			return;
